refactor: reduce file bloat
