--- pkg/config/devstack_todo.go
@@ -1,19 +1,0 @@
-package config
-
-import "os"
-
-func DevstackGetShouldPrintInfo() bool {
-	return os.Getenv("DEVSTACK_PRINT_INFO") != ""
-}
-
-func DevstackSetShouldPrintInfo() {
-	os.Setenv("DEVSTACK_PRINT_INFO", "1")
-}
-
-func DevstackEnvFile() string {
-	return os.Getenv("DEVSTACK_ENV_FILE")
-}
-
-func ShouldKeepStack() bool {
-	return os.Getenv("KEEP_STACK") != ""
-}
