--- pkg/config/docker.go
@@ -1,19 +1,0 @@
-package config
-
-import "os"
-
-type DockerCredentials struct {
-	Username string
-	Password string
-}
-
-func (d *DockerCredentials) IsValid() bool {
-	return d.Username != "" && d.Password != ""
-}
-
-func GetDockerCredentials() DockerCredentials {
-	return DockerCredentials{
-		Username: os.Getenv("DOCKER_USERNAME"),
-		Password: os.Getenv("DOCKER_PASSWORD"),
-	}
-}
