--- pkg/config/metrics.go
@@ -1,17 +1,0 @@
-package config
-
-import (
-	"github.com/spf13/viper"
-
-	"github.com/bacalhau-project/bacalhau/pkg/config/types"
-)
-
-// TODO ensure these exists or return an error
-
-func GetLibp2pTracerPath() string {
-	return viper.GetString(types.MetricsLibp2pTracerPath)
-}
-
-func GetEventTracerPath() string {
-	return viper.GetString(types.MetricsEventTracerPath)
-}
