--- pkg/config/todo.go
@@ -1,11 +1,0 @@
-package config
-
-import "os"
-
-// PreferredAddress will allow for the specificying of
-// the preferred address to listen on for cases where it
-// is not clear, or where the address does not appear when
-// using 0.0.0.0
-func PreferredAddress() string {
-	return os.Getenv("BACALHAU_PREFERRED_ADDRESS")
-}
