+++ pkg/repo/fs_test.go
@@ -1,3 +1,5 @@
+//go:build unit || !integration
+
 package repo
 
 import (
